[24-07-10] 용병 공격구현

+++ Assets/Scripts/Team/MercenaryData.cs
@@ -10,7 +10,7 @@ public class MercenaryData : ScriptableObject
     [field: SerializeField] public int EnchantLevel { get; private set; }
     [field: SerializeField] public float CoolTime { get; private set; }
     [field: SerializeField] public float Range { get; private set; }
-    [field: SerializeField] public float Damage { get; private set; }
+    [field: SerializeField] public int Damage { get; private set; }
     [field: SerializeField] public int CriticalPer { get; private set; }
     [field: SerializeField] public string Name { get; private set; }
     [field: SerializeField] public string Description { get; private set; }
